refactor(rtc_types,rtc_tenclave): move SealingError from rtc_tenclave to rtc_types::enclave_messages::errors
